data_visualization/kafka_consumer.py
import json
from kafka import KafkaConsumer

if __name__ == '__main__':
    # Kafka Consumer
    consumer = KafkaConsumer(
        'test_psd',
        bootstrap_servers='localhost:29092',
        auto_offset_reset='earliest'
    )
    for message in consumer:
        print(json.loads(message.value))
